Switch scrabble_score to direct array indexes
Instead of doing map lookups, switch the scoreSheet to be a directly
indexed array.  This means that a lookup requires just calculating the
office instead of looking up in a hahs table.  This change drops the
lookup time from about 700ns/op to about 180 ns/op, or about 3.8x
faster.